Refactor: extract duplicated swipe-row logic into shared component
